+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Enhancements
   * make unittest easier
   * add `with_exception_handlers` option to Maru.Test
+  * allow maru to ignore `MIX_ENV=test` or force `test` for other `MIX_ENV` by `config :maru, test: TRUE_OR_FALSE`
 
 ## v0.10.6 (2016-11-20)
 * Enhancements