+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 
-[Upgrade Instructions](https://maru.readme.io/docs/upgrade-instructions-from-v09) From v0.9 to v0.10
+[Upgrade Instructions](https://maru.readme.io/v0.10/docs/upgrade-instructions-from-v09-1) From v0.9 to v0.10
+[Upgrade Instructions](https://maru.readme.io/v0.11/docs/upgrade-instructions-from-v010) From v0.10 to v0.11
 
 ## v0.11.0-dev
 * Enhancements